Added clear button
This commit adds clear button in the plugin container next to the year drop down.

#Purpose
Once month picker is applied to a text field, it is not manually editable anymore. There is no way to clear the field once populated. So there should be a button to clear the value and make the field blank again.